PR comments fixed && components names changed
