+++ components/Layout.tsx
@@ -1,7 +1,6 @@
 import { PropsWithChildren } from 'react'
 import styles from '../styles/Home.module.css'
-
-import Navbar from './Navbar'
+import Navbar from './Navbar/Navbar'
 
 export default function Layout({ children }: PropsWithChildren) {
   return (
