+++ components/Navbar/Navbar.tsx
@@ -6,6 +6,7 @@ import {
   getUserLocal,
 } from '../../services/reventService'
 import Notification from '../Notifications/Notification'
+import NavigationMenu from './NavigationMenu'
 
 export default function Navbar() {
   const router = useRouter()
@@ -26,7 +27,7 @@ export default function Navbar() {
       setUser(user)
     }
     fetchUser()
-  }, user)
+  }, [user])
 
   useEffect(() => {
     async function fetchNotificationsData() {
@@ -176,63 +177,7 @@ export default function Navbar() {
               />
             </button>
             {showUserMenu && (
-              <div className='w-48 absolute right-0 mt-2 bg-white text-gray-800 rounded-lg shadow-lg p-4 max-h-60 overflow-y-auto'>
-                <ul>
-                  <button
-                    onClick={() => handleNavigation('/profile')}
-                    className='flex items-center mt-2'
-                  >
-                    <Image
-                      src='/images/profile.svg'
-                      alt='profile'
-                      width={40}
-                      height={40}
-                      className='mr-2'
-                    />
-                    <span>View Profile</span>
-                  </button>
-                  <button
-                    onClick={() => handleNavigation('/myphotos')}
-                    className='flex items-center mt-2'
-                  >
-                    <Image
-                      src='/images/myphotos.svg'
-                      alt='profile'
-                      width={40}
-                      height={40}
-                      className='mr-2'
-                    />
-                    <span>My Photos</span>
-                  </button>
-
-                  <button
-                    onClick={() => handleNavigation('/settings')}
-                    className='flex items-center mt-2'
-                  >
-                    <Image
-                      src='/images/settings.svg'
-                      alt='profile'
-                      width={40}
-                      height={40}
-                      className='mr-2'
-                    />
-                    <span>Settings</span>
-                  </button>
-                  <button
-                    onClick={() => handleNavigation('/#')}
-                    className='flex items-center mt-2'
-                  >
-                    <Image
-                      src='/images/signout.svg'
-                      alt='profile'
-                      width={40}
-                      height={40}
-                      className='mr-2'
-                    />
-                    <span>Sign Out</span>
-                  </button>
-                </ul>
-              </div>
+              <NavigationMenu setShowUserMenu={setShowUserMenu} />
             )}
           </div>
         </div>
