+++ components/NotificationsList/NotificationsList.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Notification from '../Notifications'
+import Notification from '../Notifications/Notification'
 import { Notification as NotificationType } from '../helpers/interfaces'
 
 interface NotificationList {
